Add post_install argument
With this argument set, long if statements beginning with If(IsConfigVarDefined("POST_INSTALL_SCRIPT_FILE")... etc. need not be. post_install=True checks if POST_INSTALL_SCRIPT_FILE is given, and if so, it sets output_script automatically. The behaviour is slightly different for each command. For ShellCommand, sudo is added if install script path is not defined, or defined as "". So are ignore arguments different from one class to another.